Move `commands` package to new application layer
