+++ main.go
@@ -1,8 +1,6 @@
 package main
 
-import (
-	"github.com/dietrichm/admirer/commands"
-)
+import "github.com/dietrichm/admirer/application/commands"
 
 func main() {
 	commands.Execute()
